Add basic template files and folders
Added boilerplate index.html code. This is the minimum code
requirement for a file to be correctly processed as an html file.

Added folder structure for css, js and img files. These are the only
folders that are an absolute necessity in a simple web page project
like this.

Added main, mixins, my-reset, typography and variables css files so it
is clear how css should be split into partial files rather than kept in a
single, long file

Added basic typography set up in typography.css. This sets the initial
font-size based on the assumption that the user's browser is set to
16px. There are also a couple of example media queries that change
the font size as appropriate.

Added margin, padding, box-sizing and link decoration resets to
my-reset.css. These are the ones I've figured out so far that I'm going
to need in most if not all projects. This will likely need to be added
to.

Added a placeholder variable to variables.css to show correct useage.